refactor: config package
