+++ packages/config/package.json
@@ -10,9 +10,22 @@
   "publishConfig": {
     "access": "public"
   },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs"
+    },
+    "./merge": {
+      "types": "./dist/merge-config.d.ts",
+      "require": "./dist/merge-config.js",
+      "import": "./dist/merge-config.mjs"
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
-    "build": "pnpm tsup src/index.ts src/merge.ts --format=esm,cjs --shims --dts",
-    "build-fast": "pnpm tsup src/index.ts src/merge-configs.ts --format=esm,cjs --shims --no-dts",
+    "build": "pnpm tsup src/index.ts src/merge-config.ts --format=esm,cjs --shims --dts",
+    "build-fast": "pnpm tsup src/index.ts src/merge-config.ts --format=esm,cjs --shims --no-dts",
     "dev": "pnpm build-fast --watch"
   },
   "files": [
@@ -32,18 +45,5 @@
   },
   "devDependencies": {
     "pkg-types": "1.0.3"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    },
-    "./merge": {
-      "types": "./dist/merge.d.ts",
-      "require": "./dist/merge.js",
-      "import": "./dist/merge.mjs"
-    },
-    "./package.json": "./package.json"
   }
 }
