+++ packages/config/src/merge-config.ts
