+++ packages/config/src/ts-config-paths.ts
